% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_wines}
\alias{filter_wines}
\title{Filter wines by type, country, and/or region}
\usage{
filter_wines(data, types = NULL, country = NULL, region = NULL)
}
\arguments{
\item{data}{A tibble from \code{load_wine_data()}.}

\item{types}{Optional character vector of types to keep.}

\item{country}{Optional single country to keep.}

\item{region}{Optional single region to keep.}
}
\value{
A filtered tibble.
}
\description{
Filter wines by type, country, and/or region
}
\examples{
df <- vinexplorer::load_wine_data()
whites <- vinexplorer::filter_wines(df, types = "white")
reds_and_rose <- vinexplorer::filter_wines(df, types = c("red","rose"))
rioja_all <- vinexplorer::filter_wines(df, region = "Rioja")
}
