% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the Wine Explorer Shiny app}
\usage{
run_app()
}
\value{
No return value.
}
\description{
Launch the Wine Explorer Shiny app
}
\examples{
\dontrun{
  vinexplorer::run_app()
}
}
